Address ADIv6 access ports through their debug-space bases.
ADIv6 uses aligned AP base addresses and 4 KiB register windows instead
of ADIv5 APSEL indices. Reusing an index would select the wrong
register.

Keep those selectors distinct, check them against the connected DP and
its advertised address width, and select both address words when needed.
Immediate AP reads and writes retain posted completion and existing
recovery. Queued AP operations and MEM-AP acquisition remain ADIv5-only.

+++ dap/jtag_integration_test.go
@@ -99,7 +99,7 @@ func exerciseJTAGDPMemory(t *testing.T, ctx context.Context, discovered bool) (u
 	t.Logf("AP1 IDR=%#08x", idr.Raw)
 	saved := make(map[uint8]uint32)
 	for _, addr := range []uint8{0, 4} {
-		saved[addr], err = dp.ReadRawAP(ctx, sel.Address(addr))
+		saved[addr], err = dp.ReadRawAP(ctx, sel.Address(uint16(addr)))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -123,7 +123,7 @@ func exerciseJTAGDPMemory(t *testing.T, ctx context.Context, discovered bool) (u
 		t.Fatal(err)
 	}
 	for addr, want := range saved {
-		got, err := dp.ReadRawAP(ctx, sel.Address(addr))
+		got, err := dp.ReadRawAP(ctx, sel.Address(uint16(addr)))
 		if err != nil || got != want {
 			t.Fatalf("restored AP1 register %#x = %#08x, expected %#08x: %v", addr, got, want, err)
 		}
